requirements.txt
absl-py==2.1.0
aiohttp==3.9.0b0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autopep8==2.3.1
Babel==2.15.0
beautifulsoup4==4.12.3
bert==2.2.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.1
dataclasses-json==0.6.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
elevenlabs==1.1.0
erlastic==2.0.0
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.3
flask-ngrok==0.0.25
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-speech==2.26.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.1.1
langchain==0.2.8
langchain-community==0.2.7
langchain-core==0.2.19
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langsmith==0.1.86
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.35.14
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyes==0.99.6
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyMsgBox==1.0.9
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyodbc==5.1.0
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.12.25
replicate==0.25.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
tflearn==0.5.0
threadpoolctl==3.4.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.15.2
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.2.1
uuid==1.30
virtualenv==20.25.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
